pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "cx_Freeze>=7.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tachibana-sherry-generator"
version = "1.0"
description = "シェリーちゃん可愛い大好き！"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "mouseinfo==0.1.3",
    "pillow>=10.0.0,<11",
    "psutil==7.1.3",
    "pyautogui==0.9.54",
    "pygetwindow==0.0.9",
    "pymsgbox==2.0.1",
    "pynput==1.8.1",
    "pyperclip==1.11.0",
    "pyrect==0.2.0",
    "pyscreeze>=0.1.29,<0.2",
    "pytweening==1.2.0",
    "pywin32==311",
    "six==1.17.0",
]

[tool.setuptools]
py-modules = ["Tachibana_Sherry_Generator"]
